Updated doc, fixed circular import
