+++ doc/references.md
@@ -1,0 +1,7 @@
+# References
+
+Gaudreault, S., Charron, M., Dallerit, V., & Tokman, M. (2022). High-order numerical solutions to the shallow-water equations on the rotated cubed-sphere grid. Journal of Computational Physics, 449, 110792. [https://doi.org/10.1016/j.jcp.2021.110792](https://doi.org/10.1016/j.jcp.2021.110792)
+[(PDF)](pdf/gaudreault2022_high_order_shallow_water_cubed_sphere.pdf)
+
+Charron, M., Zadra, A., & Girard, C. (2014). Four‐dimensional tensor equations for a classical fluid in an external gravitational field. Quarterly Journal of the Royal Meteorological Society, 140(680), 908–916. [https://doi.org/10.1002/qj.2185](https://doi.org/10.1002/qj.2185)
+[(PDF)](pdf/charron2013_4d_tensor.pdf)
