+++ doc/sphinx/source/doc/pdf
@@ -1,0 +1,1 @@
+/home/vma000/code/wx_factory/doc/pdf
