+++ wx_factory/pde/pde_euler_cubesphere.py
@@ -3,10 +3,10 @@ from mpi4py import MPI
 from common import Configuration
 from common.definitions import idx_rho, idx_rho_u1, idx_rho_u2, idx_rho_w, idx_rho_theta, p0, cpd, cvd, Rd, gravity
 from geometry import CubedSphere3D, Metric3DTopo
+from init.dcmip import dcmip_schar_damping
 
 from .pde import PDE
-from rhs.fluxes import rusanov_3d_hori_i_new, rusanov_3d_hori_j_new, rusanov_3d_vert_new
-from init.dcmip import dcmip_schar_damping
+from .fluxes import rusanov_3d_hori_i_new, rusanov_3d_hori_j_new, rusanov_3d_vert_new
 
 
 def compute_forcing_1(f, r, u1, u2, w, p, c01, c02, c03, c11, c12, c13, c22, c23, c33, h11, h12, h13, h22, h23, h33):
